BGDIINF_SB-2808 : switching from full timestamp to year in URL params
As there are some layers that uses the first day of the year to describe the full year, and others the last day of the year, I thought it would be easier to switch to a "only-year" notation. This also prepare the grounds for a complete switch to a more OGC compliant approach, with ISO 8601 timestamps (and ranges). Meaning when we only give a year as YYYY, we want all the data from this year (from YYYY-01-01 to YYYY-12-31)