BGDIINF_SB-2859: Removing legend button
Issue : Some layers have no legend, and upon clicking the legend button,
the legend popup shows up and looks like it's perpetually loading
content.

Fix : We added a `hasLegend` getter in the abstractLayer class, which is
then called to decide wheter we show the legend button or not.

Currently, we consider KML and external layers has having no legend. A
refinement might be needed further down in the line to ensure layers
with legends show the legend icon, but this is sufficient for now.

importing layertypes

remove logs