BGDIINF_SB-2609: consistent rounding
Issue : When displaying longitude and latitude, when the last figure of
a number was 0, we would display numbers with a different significant
number after the decimal.

Fix : We now use the "toFixed" native js function to have a fixed amount
of significant numbers after the decimal.

remove import that shouldn't be here

only calling the fixed numbers in the appropriate cases

adding rounding for all float coordinates

linting