BGDIINF_SB-2808 : add "preview year" concept to the store
As with a preview layer, this year takes precedence over any other configuration in layers. This will help when we will want to add a time slider, as we don't want the time slider to alter all active layers config, but only "preview" the year that was chosen through the time slider.
This means that all URL generation for layers now have to take into account that, if a preview year is defined, they have to replace any timestamp found in the layer's config with the timestamp corresponding to the preview year.

I've also simplified some props/data transfers between the OpenLayersInternalLayers and all specialized layers.