BGDIINF_SB-2808 : remove preview year visualization from menu
As it was confusing to show the change of preview year here, but then not highlighting the previewed year if the year selector popup was opened.
And as we want the time slider to be a "preview only" feature, it wasn't making much sense to change the year shown in the menu, as the year in the URL wasn't affected by the year selected through the time slider.