BGDIINF_SB-2808 : renaming layers time config timestamps
As it was confusing to have to write timestamp.timestamp, so I've opted for a timeEntry name that contains the year and equivalent timestamp.
In the process, I've tried to simplified the case of no time entry for a WMTS layer, and replaced all the "magic numbers" (9999, 'all', 'current') with constants that are imported when needed.

I also fixed an issue when the time slider was dealing with a WMTS layer that had a 'current' timestamp available (as it is a different value than the 'all' from WMS it was required to manage this case too)