BGDIINF_SB-2958 : fixing tests
mouse position test was failing because the viewport (the body) was stretched outside of the Cypress' iframe, because of some misinterpretation of CSS vw and vh values. So I've switch all usage of vh and vw to good old percent values. I've also grabbed the OpenLayers map specifically, as it wasn't reliable to grab the map module in this test cast.
Fixed the projection switching, that was broken due to my last refactoring.

The Cypress layer test was containing a .only, hindering other tests to be run correctly.

In the profile test, there was some misnaming of the event required to show the tooltip over the profile
fix