BGDIINF_SB-2976 : disable the 3D button if no WebGL support
also changing the 3D icon that was used, so that it adapts better to the disabled state. Before it was a PNG and it was staying red, even though the button was light grey (disabled).

Also moving the tooltip / title of the button in a popover, to better explain what this button is (there isn't any indication it is for 3D anymore) and also adding the no-WebGL text there if needed (better than having a not so easy to find HTML tooltip or an alert)