BGDIINF_SB-3008 : simplification of utils function in e2e test
removing mandatory "lang" param as the first, instead the first param of goToMapView (and children functions) is the query. Lang will be defaulted to "en" if it is not present in the query (to ensure there's always one param, and that we can then write "?" in the URL without breaking the URL scheme)

Also removing the "destructure/deconstruct" version of goToMapView, as it was used only once and was adding unwanted complexity for only one case.