BGDIINF_SB-2890 : update tests with new coord system/format approach
Changing how marker-text determines the ending coordinate of the drag and drop of icon, and text placement so that we do not use magic number anymore. Exposing the OL map instance to Cypress to achieve that.
Removing start of URLs from the 3D layers test, as it was sometimes clashing with stagging (mismatch of base URLs)