BGDIINF_SB-2890 : rework of the popup component hierarchy
removing the provide/inject approach, and use instead a slot
I had to remove the use of the "out of the box" OpenLayers Overlay, as it was creating many width issue with the content of the popover. Instead it now follows the same logic as what is done on the Cesium side.

Also fixing some issue that I detected in the CoordinateSystemBounds isInBounds and splitIfOutOfBounds, as it was not taking the border of the bound as in bounds
renaming util function reprojectUnknownSrsCoordsToWebMercator into reprojectUnknownSrsCoordsToWGS84 as it was misguiding to name the output WebMercator (usually we use this name for coordinates expressed in meters instead of degrees)