BGDIINF_SB-2890 : reproject map content when projection changes
As some data from backends are processed (re-projected), we need to reproject any data of the sort at each projection change.

Plus the map view in OpenLayers must change, as we do not have the same constraints when we use Mercator as the one with LV95.
In LV95 we want to constraint the resolutions to match exactly the ones from all our printable production (NationalMap 1:25'000, 1:50'000 and onward)