BGDIINF_SB-2854 feature : Show area and length of drawings Issue : Are and length of polygons drawn were not shown
Fix : We now check if the polygon are closed. If they are, we calculate
both area and perimeter. If not, we only calculate the perimeter. We
then show them in the popup.

Still to do :
Font Awesome Icon for filled area is not loading properly, I need to
check this