BGDIINF_SB-2890: Fixed drawing feature edit
When selecting a drawing feature outside the drawing mode (to see the details
and descriptions) and then entering the drawing mode, the feature stayed selected
but was not initialized properly to allow edition. This means that we could
edit it, but the changes were not reported to openlayer and thus not saved
on the kml, which means that the changes were lost when exiting the drawing mode.

A simple solution to this issue, is to simply deselect everything on entering
the drawing mode. This also make sense in user point of view.