Fixed time slider toggling
When only one time config layer was visible and you entered the time preview
without then changing any preview year, if you changed the year in the active
layer config the timeslider was not toggled (made invisible). This was only
the case with one layer visible at a time and when we did not changed the
preview years.

The bug was that in that case the previewYear in the store was not changed
and the menu module could not now of this change and react on it.