BGDIINF_SB-2890: Removed unnecessary wrap X coordinates in sketch profile
During the drawing of the profile we need to sketch the point of the mouse.
For this we did a wrap X coordinates, I don't know the reason why and it appears
that it is not needed as open layer can handle both coordinates (wrapped and non
wrapped), because this coordinate is only a temporary coordinate I think that is
why it is not needed.

Because to correctly wrap a coordinate you need to know in which projection you
are and that in this event type you don't know it, I removed the wrap x for
simplicity.