BGDIINF_SB-2890 : fix search tests
search now uses lat/lon from the backend, as there was some issue while using x/y with sr=2056. Might be a bug on the backend, but those coordinates were out of bound of LV95 when searching for addresses... Lat/Lon were correct, so I used them instead.
Also adapting the fixture file to have only LV95 values (as that is what we are requesting from the backend, even though x/y are somewhat wrong on the real backend)