BGDIINF_SB-2890 : change GeoJson and KML projection handling in 3D
As the Cesium logic was doing essentially the same operations as in OL, and then convert the OL object to Cesium : I've moved the reprojection logic for GeoJSON into a dedicated utils file. This can be then called once in OL and once in Cesium.
I also tried to simplify, or at least better document how this Vue mixin for Cesium primitive layer works.