BGDIINF_SB-3071: Fixed bug in import tool for extent projection
Some external WMS GetCap might list several bounding box using different projection.
That was the case with https://wms.geo.ag.ch/ where the first bounding box
in the list was CRS:84. Currently we don't support this projection which is
similar to ESPG:4326 but has an inversed axis order (lat/long instead of long/lat).

However in this use case the second bounding box was in LV95 which is supported.

So now instead of taking the first bounding box and doing a reproject, we
test simply over all bounding box and only try to re-project the one that we
support, otherwise we don't set the extent if the boundingbox is not found or
of an unsupported projection.