BGDIINF_SB-2890: Fixed legacy parameters coordinates
The legacy parameter coordinates were not correctly reprojected. This was
due to the fact that the reprojection happened in a loop several times instead
of once.

This has been solved by reorganizing a bit the code.

Also corrected a test case that did not make sense with the new lv95 default
projection. Now all legacyParamImport.cy.js test are passing.