BGDIINF_SB-3162, BGDIINF_SB-3163: Fixed background in lv95
Now the background are the regular old products, pixelkarte-grau and swissimage
in lv95 projection.

Unfortunately I could not easily add the lightbasemap and the world image
as other options in Web Mercator mode because the current BackgroundSelector
cannot be dynamic, it is made of lots of CSS magics with a hardcoded value
in css on how many background is available. So we cannot have a dynamic number
of available background without reworking the bg selector.