BGDIINF_SB-3168: Fix external layer parsing error (no `layer.name`)
Some WMS layers don't have a `layer.name` but do have a `layer.title`, like for
instance the https://bio.discomap.eea.europa.eu/arcgis/services/Ecosystem/Ecosystems/MapServer/WMSServer

In this case the import took was successful `Loading OK!` but no result
was displayed and no error was found in the error log !

Now we support also layer without name but with a title and in case the
layer has no title and no name, an error log is printed.
Also when no layers could be imported we also throw an error.