BGDIINF_SB-3154 : add e2e test to cover crosshair URL param
Handles correctly removal of crosshair param from the URL (removes the crosshair entirely)

Changing a bit how the store sync router handles interaction with the store, to make sure we wait for the dispatch to have occurred (hence the async...await instead of previous then(...))

Adding a Cypress utility (and necessary code in the app) so that Cypress can now wait for the sync between the URL and the store be finished.
Adding another Cypress utility function that changes a URL param on the fly, without reloading the app. For that I needed to expose to Cypress the VueRouter instance and its history.