BGDIINF_SB-3174 : mocking up 3D tiles when testing with Cypress
should speed up the testing, less waiting on external resources

I've added the relevant waiting function, so that we may wait for Cesium to have loaded all the tileset and requested the fake tile (this way we are 100% sure it has finished loading, and we can start interacting with the map).

There's still an issue that my fake tile is breaking the ray tracing with Cesium, I had to disable the last portion of the click test, as Cesium wasn't returning a geographic coordinate anymore. Maybe a cleverer tile / tileset coordinates could fix that later.