BGDIINF_SB-3185 : rework of the BG selector
Component based approach, so that it is possible to change the list of BG dynamically (we want to swap the BG list when going 3D).
For images, I removed the sprite approach and added one PNG per layer. The image is then loaded based on the layer ID.

I've only reproduced the "mobile" version of the BG selector, but it spreads horizontally on desktop. I find that it is better that it mimics the look and feel of the other button (zoom, geolocation, etc...)