BGDIINF_SB-3185 : squared BG selector on desktop
Add the square design back, but this time with a more component based approach, leveraging the new composable API to try to keep it as DRY as possible.

Most of the CSS was common to both style of BG selector wheel, so it has been regrouped into a single SCSS file (with mixins, akin to functions, so that it can dynamically adapt the style depending on the button size).

I've also added i18n keys based on the layer ID, instead of trying to do some kind of mapping between them and the i18n key.

The BG images have also been updated to be a bit more high rez (I tried to find a place in Switzerland without labels but a lot of things to see on a map, say hello to Münsingen)