Remove lib ogc-parser
While updating libs, I had issues with ogc-parser as it is using very outdated dependencies.
After checking the GitHub repo, I saw it was kind of staled, so I replaced it with OpenLayers function. I had to add some browser API equivalent to NodeJS in order to achieve this (see first lines of the file) as this is then used under the hood by OpenLayers to parse XMLs