BGDIINF_SB-3193 : topics are now group of layers
preparation work so that we can feed an external getCap parsing to the LayerCatalog component

Removing all related TopicTree classes, and replacing them with the GeoAdminGroupOfLayers equivalent.
Adapting the menu to use this new GeoAdminGroupOfLayers to build the topic tree.

also fix : now implementing properly the opening of certain categories, if the backend tells to do so (switch to Aviation topic to have an example)