Improve mobile navigation in 3D
I found that most (of not all) Cesium sandcastles (online examples) have this CSS rule to block touch action, it fixes all issues I experienced with panning while on a mobile device!

I also de-activated the location popup (right-click) on mobile, as it wasn't very consistent how to trigger a right click with a touch screen.