BGDIINF_SB-3109 : add menu component
First commit : we apply some small changes to the store and the compare
slider to make it easier to use the default value, as well as
deactivating the compare slider automatically if there are no more
layers.

BGDIINF_SB-3109: Adding Menu Component

Visible if there are layers
Since the default value is -0.5 for the compare slider, Toggle either
put it to -0.5 or 0.5

formatter and linter were not happy

BGDIINF_SB-3109: Compare Slider Menu

Toggle works
When there are no layers, the compare slider is automatically removed

To Fix :
Compare Menu should not be visible if there are no layers available
When toggling, doesn't cut immediately

Solving slicing on toggle

Toggling the compare Slider now correctly re-render the map with the
correct cut, unbinding or binding the layer correctly

BGDIINF_SB-3109: switch to composition API. Left the 'components' part as a comment for now, as I don't know yet what the equivalent for composition API is.

unless I'm istaken, components no longer need to be declared