Identification of KML layer features without the help of OpenLayers
so that it may work the same way when using Cesium

I had to split the deserialize function of EditableFeature in two to achieve this, as the "old" deserialize wasn't receiving JSON data directly but the OpenLayers feature. So there's now a static function for when we have an OpenLayers' feature, and one (that is used by the aforementioned one) whose task is only to transforme a JSON object into the relevant EditableFeature instance.