Wrapping all props in a toRefs call
instead of individually transforming them into refs. Sadly, it wasn't possible to do it in a one-liner, as defineProps is a pre-compiled by Vue, and if we wrap its output, the pre-compiler breaks down and can't handle it. So I added a new line with this "toRefs" call just after the defineProps

As defineProps is a compiler macro, I've also removed it from all imports (same thing for defineEmits), the vue compiler was giving us a warning about that in the console.