BGDIINF_SB-3194: Focus on provider input when opening provider list
This allow a quicker use of the providers. Note this is only done when it is
a real click and not when the user use tab to focus on the toggle button and press
enter or space to activate it. In this case the user uses the keyboard and we
should not make decision on what he want to focus.