Fixed external WMTS/WMS base url issue
Technically an external WMS/WMTS server can be served by several aliases
(different hostnames or path). In this case if the user added the layer using
the alias2 and alias1 was specified as getCap URL in the Capabilities answer
of the server, the external layer could not be loaded properly because the
alias2 was used in our URL param and alias1 was used to search for the active
layer which results in active layer not found.

To avoid such issue keep the original url as get cap url.