Changing how we trigger mouse/touch event in OL
relying on the underlying HTML element instead of the OpenLayers instance, as it was delaying some event for their own processing, breaking our code.
I've add to check that the event I receive are triggered on the canvas itself (on the map) and not elsewhere (i.e. in the popup). It would otherwise pass down the event from the popup interaction to the map (triggering a click at the "position" of  popup element)