BGDIINF_SB-3211 : fix zoom after coming back to 2D
Also removing/simplifying the geolocation tracking/untracking when the map is dragged. As we now have a central composable that handles all map interaction, this state isMapBeingDragged was not necessary anymore, it could be handle in the mouse click composable.

The zoom level was not set properly when the view (dependent on the projection) was changed in the OL composable.

I also removed the spinner backdrop from the drawing module when it closes, replacing it by a spinner inside the close button (to give the user a feedback something is happening anyway)