PB-87: Fixed legend of "Sismique réflexion" layer on mobile phone
This layer legend (as well as probably other legends) was not displayed
correctly because it contained an image that was wider than the screen. To
avoid such issue I removed the fluid mode which is not working well
with image with a fixed size.

Now this layer legend is displayed correctly with an horizontal scroll bar.