PB-87: Fixed the fuild mode of modal window (drawing share popup)
In the drawing the share popup on the mobile did not take the whole size
of the screen which looks ugly. We could either solve it by removing the
fluid mode, but then on desktop the modal was way too large, or by setting
the width to max-content. This also fixed the drawing erase modale.