PB-87: Fix search bar look and feel and its user experience
The search bar round corner were not always correct.

Also when entering for example `ber` then a short pause followed by `n`, the `n`
was sometimes automatically deleted. This was because the result of the search
triggered by a short pause (due to debouncing) reset the input value.

Now the input value is decoupled from the search query which allow a much better
user experience. This has however the drawback that when manually entering
the search param query it is then not reflected on the input ui. But this is
in theory not a use case that we should support.

The search code has been also simplified. Using the v-click-outside simplify the
code and could also solve race condition issues on cypress.