BGDIINF_SB-3153 : fit to the biggest border of the extent
calculating the corresponding zoom level on the height of the extent if it is greater than the width

Also zooming out one step after calculating the zoom level. This is required because LV95 doesn't calculated a floating zoom level, but works with steps of zooms. So the one we are provided is the closest fixed zoom level to show the extent on the whole available viewport.
We want the extent to be contained in the viewport, so we zoom out one step to achieve that.