PB-95: Added external KML support
To allow support of local KML file, the file name is used as URL and during the
parsing of the URL parameter if the url part of the layer doesn't starts with
http then it is ignored and not added to the layer list.

I also changed a bit the isLoading logic during the parsing of the url parameter for
KML. I did this for 2 reasons:
 1. Avoid unnecessary spinner which makes the application feeling to be slow
 2. In the case of an external KML layer we cannot load the metadata and it would spin forever

Due to his isLoading changes I also updated the load kml plugin to only load
kml metadata when the kml is not external and it has missing kml metadata.

Simplified KMLLayer class constructor. The fileId can always be parsed from the
URL (for non external layer). In order to simplify the constructor which has
already quite a lots of parameter remove this one and parse it from the URL
whenever possible. Same for the isLoading and isExternal, those two can be
guessed from the other parameters.