PB-95: Temporarily enable experimental defineModel feature until upgrade to vue 3.4
This feature has been adopted in Vue 3.4, therefore it is safe to use it now
within vue 3.3.

I tried to upgrade to Vue 3.4, but it breaks the drawing module ! Apparently
Vue 3.4 has a bug when returning a ref from a composable and watch this ref in
a component using the composable. In this case the watch is not fire when the
ref changes.

This was the case where in `useDrawingModeInteraction.composable.js` the returned
ref `lastFinishedFeature` did not fire change event on the watcher in `DawingMarkerInteraction.vue`