PB-97 : implement zoom to extent in 3D
The sync camera plugin was monodirectional, and wasn't reacting to changes made to the center or zoom (which is what is triggered when we do a zoomToExtent)

As this plugin is listening to mutation it triggers itself, I had to defend against infinite calls with some flagging inside the sync module.

I decided to revert the camera rotation on all axis when zooming to an extent, as I found it too difficult to calculate the equivalent without being straight into the Cesium viewer, and having access to function of the Cesium camera.