PB-95: Allow partial update of a layer
This is needed in order to speed up boot procedure. For example for the kml
layer at startup we need to load the kml and its metadata based on its
url, this could be done asynchronously and then update only the kml data of
the layer and/or only the kml metadata of the layer. Before we needed to do
both together and update them together.