PB-95: Solve cypress source maps debugging and config missmatch
In order to debug using the browser debugger tool during cypress tests you
need to have the source map and for this you need to provide the vite config
to the cypress file pre-processor.

Also we need to have the server mode and the cypress mode in sync in order
to use our environment variable in tests, otherwise the test will get the
prod variable while the server is using the dev variables.