PB-95: Solve Vue 3 reactivity issues
The Vue 3 and Vuex documentation specifies that we should only use javascript
plain object as reactive object.

Unfortunately we don't do this with our active layers that are javascript class
and put into the store. For this use case it is quite convenient and changing
this would require a big rework.

Therefore we keep the class but with a big warning. The main known issue is
that the class getter/setter cannot be reactive which leads to subtle bugs
hard to find. Reactivity could also have some issues if the class contains
side effect methods (method that set properties).

Here I removed all getter/setter and added the warning.