PB-95: add an animation when we open the import file popup
using GSAP to achieve this, as it is much easier to set multiple CSS properties this way, and it has a built-in animation engine when changing the CSS style.

I couldn't find an easy solution to place this element on top of its parent (the menu "Import file") so I resorted to place it behind the menu when it opens up.