Added some cypress data-cy
Removed in layers.cy.js a loop over menu topic catalogue items that made the
test failing with the new selectors. I did not fully understand the reason of
this loop but I guess it was kind of a work around due to selectors issues.
Now with the selector we can directly open the correct menu item without any
loop.