BGDIINF_SB-3109: compare slider menu element coherent with other advanced tools menu elements
Applied a :is-selected attribute to the compare slider menu element to
make it change color once selected.
Made a small function which returns a boolean for this, which is now
used in the toggle function too.

BGDIINF_SB-3109: fixing issue found on MacOS with slider size

The slider size and cutting part wasn't interpreted the same way on a Mac, so I used a translate-middle-x instead of the hard -20px on the cutting calculation

BGDIINF_SB-3109: small correction to compare slider max bounds