PB-246: Fixing legacy params for timestamps
Issue : The timestamps given as a legacy parameter are not taken into
account, and time enabled layers always default back to their base
value.

Fix : Issue was caused by a small typographical error, which stopped
the parameters to be taken into account. Also, the function which sets
the year for a layer wasn't working (vuex enforcement to avoid changes
outside of store)

Further Fix needed: While this now works for the Journey through time,
it doesn't work for the swissimage (ch.swisstopo.swissimage-product),
with a valid year. We're

went around the fact that some layers might have years already instead of timestamps, and made sure to pick the one that's defined