PB-88: Fix checkOlLayer cypress command bug
The `args` `visible` and `opacity` of previous call were kept due to argument
being passed as reference, therefore if the args is an object or an array
we need to copy it.

The other issue was that the visible and opacity are boolean or number and
that visible = false and opacity = 0 where overwritten with the default value.

The last issue was that when a layer is visible = false it should not be added
at all to open layer map !