PB-246: Handling legacy timeseries
Issue : When setting legacy parameter for layers, the timeseries
required to go through a mutation to set it. The problem was that the
setter checked to add these only to an active layer.
Since the legacy parameter handler assigned values to a layer that might
not yet be active, this caused issues with the parameter not being able
to update.
Solution : We create a new setter which sets directly the timeserie to a
layer.