PB-246: regex fix and timeConfig clones implementation
The regex in the legacy parser recognised single layers without
parameters as layers with new parameters. It also flagged any layer with
a `-` as a new layer. This has been corrected.
The timeconfig object within time-enabled layers is now deep cloned too
when calling the layer clone() method, to allow the modification of
cloned layers timeConfigs outside vue mutations.