PB-246: legacy parameter parser update
Divided the legacy parameter parser in two entities : one that checks if
the search string is a legacy one, and one that parses it.

Moved both functions to legacyLayerParamUtils, as they would be needed
in more than one file, and it made more sense to have them as a tool
than a part of the router plugin.

Changed the topics calls to the legacy parameter handler for layers, to
be the same as in the rest of the code

Small security in legacy parameter checker

In topics, the 'search' query we expect might sometimes be null, thus a
small additional check to ensure we handle these correctly

modifying the Url parsing, using URLSearchParameters objects instead

lint

remove parseLegacyParams function