PB-246 : withHash default value is true in tests
the `withHash` parameter define if the query is sent with the legacy
parameters (false) or the modern parameters (true)
Upon changing the way the legacy parser works, we also removed the
possibility for modern parameters to be correctly interpreted within the
legacy parser.
The issue is : most tests never set this withHash parameter, and thus
were sending modern parameters through the legacy parser.
Now, we will send everything through the correct handler, and only use
the legacy handler for testing the legacy handler.

Using correctly the foreach of URLSearchParameters. Also added the non-default 'true' withHash parameter to legacy e2e tests