PB-114: Fix KML feature description
In KML spec a Feature description should be inside the description tag within
the placemark. This was not the case.

To do so using the openlayer library the description needs to be part of the
feature property, therefore add export the description into the feature before
exporting the features into a KML string.