PB-269 : better selection of drawing feature (post-drawing)
We return a LayerFeature instead of the specialized one for drawing, making our feature identification compatible with external KML in the process.

Also changing how the popup for KML feature is built, and using the layer title (common practice between layers) and trying to find a description for this specific feature. If no description is given, fallback to a "no further information" text instead of leaving the popup empty.