PB-190: Implemnting bod layer Id
Issue : We want to implement the {bod-layer-id} parameter functionality
as follows :

parameter 1: {layer-id}={features,id,separated,by,commas}
	This should activate the layer, and select the features
parameter 2: showToolTip={true, false}
	If set to false, the features should be highlighted, but the
tooltip should remain hidden. This is only valid at startup.

WIP

PB-190: Router approach to the implementation

We added a small router that should be called after the legacy
Parameters router, whose goal is to get rid of the dynamic parameter in
the URL, transfer it to the 'layers' param and ensure it is visible.