PB-259: Improve topic changes performance and plConfig layers order
When changing the topic we need to set clear the active layers and set a bunch
of new activated layers. This was done with several dispatch which was quite
inefficient.

Now we use only one dispatch with the new array, this way for the default topic
`ech` we gain about 500ms on development build.

Also fixed a bug in the infamous plConfig of topics (used by schneesport topic)
in which layers were added in reverse order.