PB-285: Added `dispatcher` to all store actions related to the URL parsing
This should also be done for all actions, but for the time being only added
it to the url parameter related actions.

This new generic parameter gives the possibility to do different actions in the
store plugins depending on the source of the event. This avoid bulky/flaky tests
using local closure flags.