PB-259: Improve language changes performance
Vuex mutation are costly and sending in a loop a lots of event is not a good
idea. When changing the languages we need to reload the layer config and
reset all active layers in order to have them translated instead of doing this
in a loop we all set them in one event.
This improve greatly performance with long active layer list.