Uses e2e best practice in drawing test
On the CI this E2E test is sometimes failing in the icon size check in KML
payload. Although this commit should not fix this issue, it removes some
bad practice of `force: true` flag and make sure that element are visible before
clicking which is a good practice (avoiding clicking on element not yet fully
rendered and avoiding race condition).