Fix cypress e2e race condition in drawing icon size test
The icon size test did failed time to time. This was due to a race condition.

The icon scale in KML is set by openlayer and uses a scale factor based on the
icon size. Now if the icon has not yet be loaded during the KML write, openlayer
takes a default size of 64 pixel which defer from our icon size of 48 pixel.
Due to this the icon scale of a large icon was 2 instead of 1.5.

We solve this race condition by waiting on all icons being loaded before doing
the icon size changes.

NOTE this is not an issue in real life as the save drawing is always delayed
at least 2 seconds which let plenty of time to load the icons. In cypress this
delay was reduced to speed up testing.