Fix race condition when exporting/saving kml
To export a kml, openlayer requires the icon size in order to compute the
scale. If the icon has not yet been loaded when generating the style then
the icon scale would be computed with a default size of 64px.

The last commit tried to solve this issue in cypress test, but unfortunately
waiting on the icon to be loaded did not always solved the race condition.

I could not find a better way to ensure this as settings the size when
generating the icon style.

This solve temporarly the issue, but in the future we should get the size
from the backend with the others icon metadata.