PB-293: Fixing E2E test
Changed the order of the expected query parameter in legacy test.

In external wmts layer, somehow the Get capabilites are called twice at startup
which made the test failing because the second wait on capabilities when
reloading did not wait due to previous request.

In a separate PR we should investigate why the get capabilities is called twice.