PB-322: Avoid reloading external layer upon browser history navigation
When changing the config (opacity, visibility) of an external layer (KML, GPX,
WMS or WMTS) and then use the browser history navigation to revert the changes,
it triggered a reload of the layer and set its name back to the default until
the layer has been reloaded.

We can simply avoid this by updating the current layer object instead of
creating a new one.