PB.2559: Fix toggling in 3d race condition
When loading the menu as async componenent, it adds a kind of race condition
when toggling in 3d mode for the first time making the application crash.

So avoid this race condition by keeping the menu module within the index
chunk.

NOTE: I did not track down the race condition to understand what was really
the issue.PB.2559: Fix toggling in 3d race condition