PB-339: Fix layer z-index computation
The logic has two issues, one is that for group of layers we counted always
one index too much, and that for the on top of all layers we did not took the
group of layers into account.

This solve the issues in the e2e tests for the external layers.