Fix save drawing race condition
When exiting the drawing, we could have a race condition between the closing
save and pending debounced save. Sometimes a pending debounce save, not started
was started after the closing and closing save, which saved an empty kml because
the drawing layer would have been removed.

This was making some test failing time to time and could also have been an
issue for a real user if he would close the drawing too quickly.