PB-341: Fix drawing e2e race condition
Apparently the debouncing of the feature title and description in drawing together
with the save drawing debouncing creates a race condition when editing the title
and/or description and closing the drawing mode. In this case the save kml
triggered by the title/description update debounced happens only when the
drawing mode is closed and when the drawing layer has removed all its
feature which result to saving an empty drawing, clearing the existing drawing.

To avoid this we speed up debouncing, this is not a perfect fix but it works
and I could not find a way of correctly cleaning up the debouncing for
the closing drawing menu event.