PB-317: Moved the tool box to the map module
This is a preparation work to allow to add the basic map tools to the embed
vue, therefore remove it from the menu module and put it to the map module.

Those tools are anyway more some basic map tools that should be part of the map
module, and we keep the menu module for more advanced stuff.

Also made the map tool box with slot extensible.