PB-294: Avoid reloading the application when entering ENTER in import file
In import file when entering enter with an invalid url field, it triggered a
page reload. This was due to the fact that the `loadFile()` was triggered with
an invalid url which triggered a fetch to the current host, reloading the app.

Now we prevent calling loadFile() if the url is not valid.