PB-332: Fixed compare slider with duplicate layers and also visibility flag bugs
Because the layers can be duplicate we need to take the one on top of the open layer
list, not that this list is by default not by zindex sorted.

The compare slice needs also to be done each time any layers from the list
changes its visibility otherwise the compare ratio is not applied to the correct
layer.